+++ src/multi_vector_simulator/utils/constants_json_strings.py
@@ -110,6 +110,8 @@ STORAGE_CAPACITY = "storage capacity"
 SOC_INITIAL = "soc_initial"
 SOC_MAX = "soc_max"
 SOC_MIN = "soc_min"
+THERM_LOSSES_REL = "fixed_thermal_losses_relative"
+THERM_LOSSES_ABS = "fixed_thermal_losses_absolute"
 
 # Constraints
 CONSTRAINTS = "constraints"